requirements.txt
# ============================================================
# Semantic Search System — Python Dependencies
# Compatible with Python 3.10+
# Install: pip install -r requirements.txt
# ============================================================

# ── Core API ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# ── Embeddings ───────────────────────────────────────────────
sentence-transformers>=3.0.1
torch>=2.3.0              # CPU-only; replace with torch-cuda for GPU
transformers>=4.41.1

# ── Vector Database ──────────────────────────────────────────
faiss-cpu>=1.8.0          # Use faiss-gpu for CUDA environments

# ── Clustering & ML ──────────────────────────────────────────
scikit-learn>=1.4.2
numpy>=1.26.4
scipy>=1.13.0

# ── Data Processing ──────────────────────────────────────────
pandas>=2.2.2
nltk>=3.8.1

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
tqdm==4.66.4

# ── Testing ──────────────────────────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0             # For FastAPI TestClient

# ── Visualisation (optional, for experiments) ────────────────
matplotlib==3.9.0
